refactor(frontend): extract periodsettings recoil functions into seperate file
